mirror: allow updating blacklisted ids
Now the GitHub mirror will warn when the blacklisted ids are
inconsistent, rather than throwing an error.

This bit of tolerance is quite convenient, because you might know that
the change in blacklisted ids was for a totally different repo, and have
a reason to want to keep your cache around. Of course, you'll eventually
want to reset the cache so that you stop seeing the warning message. :)

Test plan: Tests updated; `yarn test` passes.