load sets plugins appropriately
The `cli/load` command now chooses plugins based on the project it is
loading.

Test plan: No automated testing, as I'm planning to rewrite this portion
of the codebase soon. I did manually test that loading project specs
with GitHub or Discourse projects produces the right plugins in scope in
the UI.