build_static_site supports loading project files
This commit modifies build_static_site.sh so that it admits loading
custom project.json files, as generated by the `gen-project` command.

This is a temporary improvement, meant to facilitate the launch of phase
1 of the CredSperiment. I expect to deprecate this code soon, when I
refactor SourceCred to have a declarative/instance based configuration.

Test plan: I've manually tested it on a simple example. I am planning to
deprecate this entire script soon, and the automated testing is a pain,
so I haven't updated the test files.