Add to (and update) Discourse snapshots (#1422)
I made a new [test post][1] which has references. The Discourse
snapshots now include it, so we can give a realistic test of reference
and mention detection.

This will allow us to verify whether #1420 is affecting us, and fix it
if so.

Test plan: Commit was generated by running the snapshot updater. Other
snapshots have been updated and look OK. `yarn test` passes.

[1]: https://sourcecred-test.discourse.group/t/a-post-with-references/21