Add non-interactive `sourcecred init` command
The `sourcecred init` command sets up a `sourcecred.json`
file, which is currently just a serialized `Project`. This is based on
work in #1496, but is non-interactive instead.

This is a first step towards implementing the [instance system].

[instance system]: https://discourse.sourcecred.io/t/sourcecred-instance-system/244

Test plan:

This command is thoroughly tested by sharness tests; see
`sharness/test_cli_init.t`.

THanks to @wchargin for help with the sharness testing.